requirement.txt
mistralai
python-dotenv
requests
streamlit
PyQt5
flask